Restructured modules.
