+++ macro_utilities/stringify_attr/attr_derive/src/parse.rs
@@ -43,11 +43,7 @@ pub fn parse_attr_input(ctx: &mut Context, meta: NestedMeta) -> Result<Vec<Meta>
         n => {
             ctx.error_spanned_by(
                 n,
-                format!(
-                    "{}. Found `{}`",
-                    EXPECTED_LIST,
-                    meta.to_token_stream()
-                ),
+                format!("{}. Found `{}`", EXPECTED_LIST, meta.to_token_stream()),
             );
             Err(())
         }
