+++ runtime/swim_runtime/src/pressure/recon/tests.rs
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use super::MapOperationReconEncoder;
 use bytes::{Bytes, BytesMut};
 use swim_api::protocol::map::{MapOperation, RawMapOperation};
 use tokio_util::codec::Encoder;
-use super::MapOperationReconEncoder;
 
 #[test]
 fn recon_encode_clear() {
